refactor(#590): extract SpectrumPanel inline logic to spectrum-logic.ts
Move 6 pure functions (parseScopeFrame, formatFreqOffset, deriveFreqTicks,
getDragInterval, isFixedScope, freqFromPixel) and ScopeFrame type to a
dedicated module. Tests now import from the shared module instead of
re-implementing the functions.
