+++ frontend/src/components/spectrum/__tests__/spectrum-panel-logic.test.ts
@@ -3,28 +3,16 @@
  * scale ticks, drag rate limiting, scope mode classification, click-to-tune.
  */
 import { describe, it, expect } from 'vitest';
+import {
+  parseScopeFrame,
+  formatFreqOffset,
+  deriveFreqTicks,
+  getDragInterval,
+  isFixedScope,
+  freqFromPixel,
+} from '../spectrum-logic';
 
-// ── Re-implemented pure functions from SpectrumPanel.svelte ─────────────────
-
-interface ScopeFrame {
-  receiver: number;
-  mode: number;
-  startFreq: number;
-  endFreq: number;
-  pixels: Uint8Array;
-}
-
-function parseScopeFrame(buf: ArrayBuffer): ScopeFrame | null {
-  const view = new DataView(buf);
-  if (view.byteLength < 16 || view.getUint8(0) !== 0x01) return null;
-  const receiver = view.getUint8(1);
-  const mode = view.getUint8(2);
-  const startFreq = view.getUint32(3, true);
-  const endFreq = view.getUint32(7, true);
-  const pixelCount = view.getUint16(14, true);
-  if (16 + pixelCount > view.byteLength) return null;
-  return { receiver, mode, startFreq, endFreq, pixels: new Uint8Array(buf, 16, pixelCount) };
-}
+// ── Test helpers ────────────────────────────────────────────────────────────
 
 function encodeFrame(rx: number, mode: number, sf: number, ef: number, px: Uint8Array): ArrayBuffer {
   const buf = new ArrayBuffer(16 + px.length);
@@ -37,36 +25,6 @@ function encodeFrame(rx: number, mode: number, sf: number, ef: number, px: Uint8
   v.setUint16(14, px.length, true);
   new Uint8Array(buf, 16).set(px);
   return buf;
-}
-
-function formatFreqOffset(hz: number): string {
-  if (hz === 0) return '0';
-  const absHz = Math.abs(hz);
-  const sign = hz < 0 ? '-' : '+';
-  if (absHz >= 1e6) return `${sign}${(absHz / 1e6).toFixed(1)}M`;
-  if (absHz >= 1e3) return `${sign}${(absHz / 1e3).toFixed(0)}k`;
-  return `${sign}${absHz}`;
-}
-
-function deriveFreqTicks(spanHz: number): { position: number; label: string }[] {
-  if (spanHz <= 0) return [];
-  return [-1, -0.5, 0, 0.5, 1].map((ratio) => ({
-    position: (ratio + 1) * 50,
-    label: formatFreqOffset((spanHz * ratio) / 2),
-  }));
-}
-
-function getDragInterval(speed: number): number {
-  if (speed > 600) return 700;
-  if (speed > 200) return 400;
-  return 200;
-}
-
-function isFixedScope(mode: number): boolean { return mode === 1 || mode === 3; }
-
-function freqFromPixel(clientX: number, rectLeft: number, rectWidth: number,
-  startFreq: number, spanHz: number): number {
-  return startFreq + ((clientX - rectLeft) / rectWidth) * spanHz;
 }
 
 // ── Tests ───────────────────────────────────────────────────────────────────
